requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
django-allauth==0.50.0
djangorestframework>=3.13.1
dj-rest-auth==2.2.4